Only restorecon /data/data if passed an explicit flag.
As discussed in the comments of:
https://android-review.googlesource.com/#/c/81292/
we do not want restorecon_recursive /data by init.rc to
change the contexts of /data/data or /data/users/N because
those contexts are dependent on package information in
/data/system/packages.list that may not be set correctly
at boot on an upgrade.  Therefore we skip /data/data
on a recursive restorecon unless passed an explicit flag.
This flag will be used by the PMS or installd in a subsequent
change to restorecon /data/data after computing the seinfo
values for packages and updating packages.list accordingly.
It will also be optionally used by the toolbox restorecon
command if passed a new option to be introduced in a subsequent
change.

Change-Id: I137588013ed1750315702c0dbe088ce3e4a29e83
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>