libselinux:  fix policy reload logic
I5d6d6eb7438361bbb072540c96361cef95c83a9e introduced a bug
in the policy reload logic such that we incorrectly (but harmlessly)
load policy twice from / on each boot and never load policy from
/data/security/current even if it is present.

Also, even prior to that change, we were failing to reload policy
from / if we previously had loaded a policy from /data/security/current
and that policy was removed.

Fix the bugs.  Also correct the comments to drop the obsolete
reference to safe mode and to reflect the updated code.

Change-Id: I7b53c91c5681764009de453ff104a72cd26d7c2e
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>